commit #7. 1. Fixed code for show search. Instead of searching for a single show, search for all shows by keywords is implemented. 2. Optimized date object classes
