--- app/src/main/java/com/pidzama/moviefind/data/model/cast/Country.kt
@@ -1,7 +1,0 @@
-package com.pidzama.moviefind.data.model.cast
-
-data class Country(
-    val code: String,
-    val name: String?,
-    val timezone: String
-)
