--- app/src/main/java/com/pidzama/moviefind/data/model/cast/Links.kt
@@ -1,5 +1,0 @@
-package com.pidzama.moviefind.data.model.cast
-
-data class Links(
-    val self: Self
-)
