--- app/src/main/java/com/pidzama/moviefind/data/model/episodes/Links.kt
@@ -1,6 +1,0 @@
-package com.pidzama.moviefind.data.model.episodes
-
-data class Links(
-    val self: Self,
-    val show: Show
-)
