--- app/src/main/java/com/pidzama/moviefind/data/model/episodes/Show.kt
@@ -1,5 +1,0 @@
-package com.pidzama.moviefind.data.model.episodes
-
-data class Show(
-    val href: String
-)
