--- app/src/main/java/com/pidzama/moviefind/data/model/movies/Image.kt
@@ -1,10 +1,0 @@
-package com.pidzama.moviefind.data.model.movies
-
-import androidx.room.ColumnInfo
-
-data class Image(
-    @ColumnInfo(name = "medium")
-    val medium: String?,
-    @ColumnInfo(name = "original")
-    val original: String?
-)
