+++ app/src/main/java/com/pidzama/moviefind/data/model/movies/Movie.kt
@@ -56,3 +56,70 @@ data class Movie(
     @ColumnInfo(name = "isFavorite")
     var isFavorite: Boolean = false
 ) : Serializable
+
+data class Country(
+    @ColumnInfo(name = "code")
+    val code: String,
+    @ColumnInfo(name = "name")
+    val name: String,
+    @ColumnInfo(name = "timezone")
+    val timezone: String
+)
+
+data class Externals(
+    @ColumnInfo(name ="imdb")
+    val imdb: String,
+    @ColumnInfo(name ="thetvdb")
+    val thetvdb: Int,
+    @ColumnInfo(name ="tvrage")
+    val tvrage: Int
+)
+
+data class Image(
+    @ColumnInfo(name = "medium")
+    val medium: String?,
+    @ColumnInfo(name = "original")
+    val original: String?
+)
+
+@Entity
+data class Links(
+    @Embedded(prefix = "previousepisode")
+    val previousepisode: Previousepisode,
+    @Embedded(prefix = "self")
+    val self: Self
+)
+
+data class Network(
+    @Embedded(prefix = "country")
+    val country: Country,
+    @ColumnInfo(name = "id")
+    val id: Int,
+    @ColumnInfo(name = "name")
+    val name: String,
+    @ColumnInfo(name = "officialSite")
+    val officialSite: String
+)
+
+data class Previousepisode(
+    @ColumnInfo(name = "href")
+    val href: String
+)
+
+data class Rating(
+    @ColumnInfo(name = "average")
+    val average: Double
+)
+
+data class Schedule(
+    @ColumnInfo(name = "days")
+    @TypeConverters
+    val days: List<String>,
+    @ColumnInfo(name = "time")
+    val time: String
+)
+
+data class Self(
+    @ColumnInfo(name = "href")
+    val href: String
+)
