--- app/src/main/java/com/pidzama/moviefind/data/model/movies/Rating.kt
@@ -1,8 +1,0 @@
-package com.pidzama.moviefind.data.model.movies
-
-import androidx.room.ColumnInfo
-
-data class Rating(
-    @ColumnInfo(name = "average")
-    val average: Double
-)
