--- app/src/main/java/com/pidzama/moviefind/data/model/movies/Self.kt
@@ -1,8 +1,0 @@
-package com.pidzama.moviefind.data.model.movies
-
-import androidx.room.ColumnInfo
-
-data class Self(
-    @ColumnInfo(name = "href")
-    val href: String
-)
