--- app/src/main/java/com/pidzama/moviefind/data/model/search/Externals.kt
@@ -1,7 +1,0 @@
-package com.pidzama.moviefind.data.model.search
-
-data class Externals(
-    val imdb: String,
-    val thetvdb: Int,
-    val tvrage: Int
-)
