--- app/src/main/java/com/pidzama/moviefind/data/model/search/Image.kt
@@ -1,6 +1,0 @@
-package com.pidzama.moviefind.data.model.search
-
-data class Image(
-    val medium: String,
-    val original: String
-)
