--- app/src/main/java/com/pidzama/moviefind/data/model/search/Links.kt
@@ -1,6 +1,0 @@
-package com.pidzama.moviefind.data.model.search
-
-data class Links(
-    val previousepisode: Previousepisode,
-    val self: Self
-)
