--- app/src/main/java/com/pidzama/moviefind/data/model/search/Network.kt
@@ -1,8 +1,0 @@
-package com.pidzama.moviefind.data.model.search
-
-data class Network(
-    val country: Country,
-    val id: Int,
-    val name: String,
-    val officialSite: String
-)
