--- app/src/main/java/com/pidzama/moviefind/data/model/search/Previousepisode.kt
@@ -1,5 +1,0 @@
-package com.pidzama.moviefind.data.model.search
-
-data class Previousepisode(
-    val href: String
-)
