--- app/src/main/java/com/pidzama/moviefind/data/model/search/Rating.kt
@@ -1,5 +1,0 @@
-package com.pidzama.moviefind.data.model.search
-
-data class Rating(
-    val average: Double
-)
