--- app/src/main/java/com/pidzama/moviefind/data/model/search/Schedule.kt
@@ -1,6 +1,0 @@
-package com.pidzama.moviefind.data.model.search
-
-data class Schedule(
-    val days: List<String>,
-    val time: String
-)
