--- app/src/main/java/com/pidzama/moviefind/data/model/search/SearchItem.kt
@@ -1,6 +1,0 @@
-package com.pidzama.moviefind.data.model.search
-
-data class SearchItem(
-    val score: Double,
-    val show: Show
-)
