--- app/src/main/java/com/pidzama/moviefind/data/model/search/Self.kt
@@ -1,5 +1,0 @@
-package com.pidzama.moviefind.data.model.search
-
-data class Self(
-    val href: String
-)
