--- app/src/main/java/com/pidzama/moviefind/data/model/search/WebChannel.kt
@@ -1,8 +1,0 @@
-package com.pidzama.moviefind.data.model.search
-
-data class WebChannel(
-    val country: Any,
-    val id: Int,
-    val name: String,
-    val officialSite: String
-)
