--- app/src/main/java/com/pidzama/moviefind/data/model/seasons/Country.kt
@@ -1,9 +1,0 @@
-package com.pidzama.moviefind.data.model.seasons
-
-import java.io.Serializable
-
-data class Country(
-    val code: String,
-    val name: String,
-    val timezone: String
-)
