--- app/src/main/java/com/pidzama/moviefind/data/model/seasons/Image.kt
@@ -1,6 +1,0 @@
-package com.pidzama.moviefind.data.model.seasons
-
-data class Image(
-    var medium: String?,
-    val original: String?
-)
