--- app/src/main/java/com/pidzama/moviefind/data/model/seasons/Links.kt
@@ -1,7 +1,0 @@
-package com.pidzama.moviefind.data.model.seasons
-
-import java.io.Serializable
-
-data class Links(
-    val self: Self
-)
