--- app/src/main/java/com/pidzama/moviefind/data/model/seasons/Network.kt
@@ -1,10 +1,0 @@
-package com.pidzama.moviefind.data.model.seasons
-
-import java.io.Serializable
-
-data class Network(
-    val country: Country,
-    val id: Int,
-    val name: String,
-    val officialSite: String
-)
