--- app/src/main/java/com/pidzama/moviefind/data/model/seasons/SeasonInfo.kt
@@ -1,6 +1,0 @@
-package com.pidzama.moviefind.data.model.seasons
-
-data class SeasonInfo (
-    val number: Int,
-    val image: String,
-)
