+++ app/src/main/java/com/pidzama/moviefind/data/model/seasons/SeasonsItem.kt
@@ -16,3 +16,29 @@ data class SeasonsItem(
     val url: String,
     val webChannel: Any
 ): Serializable
+
+data class Country(
+    val code: String,
+    val name: String,
+    val timezone: String
+)
+
+data class Image(
+    var medium: String?,
+    val original: String?
+)
+
+data class Links(
+    val self: Self
+)
+
+data class Network(
+    val country: Country,
+    val id: Int,
+    val name: String,
+    val officialSite: String
+)
+
+data class Self(
+    val href: String
+)
