--- app/src/main/java/com/pidzama/moviefind/data/model/seasons/Self.kt
@@ -1,6 +1,0 @@
-package com.pidzama.moviefind.data.model.seasons
-
-
-data class Self(
-    val href: String
-)
