+++ app/src/main/java/com/pidzama/moviefind/ui/main/search/SearchFragment.kt
@@ -8,8 +8,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
+import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.pidzama.moviefind.data.model.movies.Movie
+import com.pidzama.moviefind.data.model.search.SearchItem
+import com.pidzama.moviefind.data.model.search.Show
 import com.pidzama.moviefind.databinding.FragmentSearchBinding
 import com.pidzama.moviefind.ui.main.search.adapter.SearchAdapter
 import dagger.hilt.android.AndroidEntryPoint
@@ -65,26 +68,26 @@ class SearchFragment : Fragment() {
         }
     }
 
-    private fun setListSearchMovie(searchMovie: Movie) {
+    private fun setListSearchMovie(searchMovie: ArrayList<SearchItem>) {
         binding.searchRecycler.run {
             if (adapter == null) {
                 adapter = SearchAdapter {
                     findNavController().navigate(
                         SearchFragmentDirections
                             .actionSearchFragmentToDetailsFragment(
-                                it.name,
-                                it.id,
-                                it.rating.average.toLong(),
-                                it.summary,
-                                it.premiered,
-                                arrayOf(it.genres.take(2).toString()),
-                                it.officialSite,
-                                it.url,
-                                it.status
+                                it.show.name,
+                                it.show.id,
+                                it.show.rating.average.toLong(),
+                                it.show.summary,
+                                it.show.premiered,
+                                arrayOf(it.show.genres.take(2).toString()),
+                                it.show.officialSite,
+                                it.show.url,
+                                it.show.status
                             )
                     )
                 }
-                layoutManager = LinearLayoutManager(requireContext())
+                layoutManager = GridLayoutManager(requireContext(), 2)
             }
             (adapter as? SearchAdapter)?.setList(searchMovie)
         }
