+++ app/src/main/res/layout/fragment_search.xml
@@ -64,6 +64,7 @@
         android:layout_marginStart="@dimen/dp_20"
         android:layout_marginTop="@dimen/dp_40"
         android:layout_marginEnd="@dimen/dp_20"
+        android:paddingBottom="@dimen/dp_160"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/inputSearch"
